chore: resolve pr comments
